Split DescriptionEntry tests into smaller modules
